third_party/blink/web_tests/wpt_internal/prerender/restriction-focus.html

<!DOCTYPE html>
<title>Prerendering documents are not focused</title>
<script src="/common/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<input type="text" id = "prerenderTextField">
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  document.getElementById('prerenderTextField').focus();
  assert_true(
      document.hasFocus(),
      'Initial document should have focus.');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {once:true});
  });

  // Start prerendering a page that accesses the focus state of the page.
  startPrerendering(`resources/prerender-focus.html`);

  const result = await gotMessage;
  assert_false(result, 'Prerendering document should not have focus.');

  bc.close();
}, 'Prerendering document should not have focus.');

</script>
</body>
